06-07-2019

+++ database/migrations/2019_06_18_114056_create_editorials_table.php
@@ -14,7 +14,7 @@ class CreateEditorialsTable extends Migration
     public function up()
     {
         Schema::create('editorials', function (Blueprint $table) {
-            $table->bigIncrements('id')->autoIncrement();
+            $table->integer('id')->autoIncrement();
             $table->text('nombre_editorial');
             $table->timestamps();
             // $table->foreign('usuario_rut')->references('rut')->on('users');
